cloud_implemetation/requirements.txt
torch
torchvision
ultralytics
ipykernel
opencv-python
python-dotenv
openai
AWSIoTPythonSDK
streamlit